requirements-dev.txt
-r requirements.txt
pytest==9.0.2
pytest-django==4.12.0
ruff==0.15.9
